Cottonee Cove Server v0.2.6j - Adding additional check in build process

+++ tools/build-utils.js
@@ -4,8 +4,10 @@ const fs = require("fs");
 const child_process = require("child_process");
 const esbuild = require('esbuild');
 
-for (const folder of fs.readdirSync('./dist')) {
-	fs.rmSync(`./dist/${folder}`, { recursive: true });
+if (fs.existsSync('./dist')) {
+	for (const folder of fs.readdirSync('./dist')) {
+		fs.rmSync(`./dist/${folder}`, { recursive: true });
+	}
 }
 
 const copyOverDataJSON = (file = 'data') => {
